backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
langgraph==0.2.39
langchain==0.3.3
langchain-openai==0.2.2
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
